requirements.txt
python-Levenshtein >= 0.27.1
python-magic >= 0.4.27
pysam >= 0.23.0